points

+++ app/controllers/points_controller.rb
@@ -25,7 +25,7 @@ class PointsController < ApplicationController
 
   def update
     point = Point.find(params[:id])
-    # authorize point
+    authorize point
     if point.update(point_params)
       render json: point, status: 200
     else
@@ -34,18 +34,34 @@ class PointsController < ApplicationController
   end
 
   def destroy
-    point = POint.find(params[:id])
-    # authorize point
+    point = Point.find(params[:id])
+    authorize point
     point.destroy
     render json: {message: "Record Destroyed Successfully"}
   end
+
+  def manager_score
+    point = Point.find(params[:id])
+    authorize point
+    if point.update(manager_score_params)
+      render json: point, status: 200
+    else
+      render json: {message: "Points Cannot be created", error: point.errors.full_messages}
+    end
+  end
+
 
 
 
   private
 
   def point_params
-    params.require(:point).permit(:score, :month, :employee_id)
+    params.require(:point).permit(:self_score, :month, :employee_id, :manager_id)
   end
+
+  def manager_score_params
+    params.require(:point).permit(:manager_score, :manager_id)
+  end
+
 
 end
